feat: Symptoms Tracker

See merge request uplift-code-camp/students/batch-24/p5-group-1/elderly_care_app!22

+++ client/src/components/LoginForm.jsx
@@ -16,8 +16,11 @@ export default function LoginForm({onLogin, setIsLogin, navigate}) {
             });
             console.log('Login successfully:', response.data)
             const token = response.data.token || response.data.accessToken;
-            if (token) { //keeps you logged in after refresh
+            const user = response.data.user;
+
+            if (token && user) { //keeps you logged in after refresh
               localStorage.setItem('userToken', token);
+              localStorage.setItem('user', JSON.stringify(user))
               onLogin();
               navigate('/home');
             }
